requirements.txt
amqp==2.5.0
backports.csv==1.0.7
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.8
django-crispy-forms==1.7.2
django-filter==2.1.0
django-grappelli==2.13.1
django-import-export==1.2.0
django-simple-history==2.7.2
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
kombu==4.6.1
odfpy==1.4.0
openpyxl==2.6.2
Pillow==6.0.0
pkg-resources==0.0.0
psycopg2==2.8.3
pytz==2019.1
PyYAML==5.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
urllib3==1.25.3
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
